feat: update pgvector, pgvecto-rs and vchord versions (#83)
* feat: update pgvector, pgvecto-rs and vchord versions

Updates the versions of `pgvector`, `pgvecto-rs`, and `vchord` to their
latest available releases. This includes copying necessary files from
the `pgvector` image to the Bitnami PostgreSQL image for extension
functionality.

- Updated `PGVECTORS_TAG` and introduced `PGVECTOR_TAG` for version control.
- Copied `vector.so` and extension files from `pgvector` image.

* feat(ci): add goss checks for vector extension and database creation

This commit adds new goss checks to verify:

- The existence of the `vector.control` file.
- Successful creation of a 'test' database.
- Successful activation of the `vchord` extension within the 'test' database.

These checks ensure the vector extension and database setup are functioning correctly during CI.

* feat(ci): specify postgres port in goss tests

This commit explicitly defines the PostgreSQL port (5432) in the `goss.yaml` file for the CI tests.

This ensures that the tests connect to the correct port, especially in
environments where the default port might be different.  The change
affects the database creation and extension activation commands.

* feat(postgres): removed database and vchord extension checks

This commit removes checks to the Goss configuration to verify, now we
are relying on path based tests.